On Linux, __CFTSRRate was never getting initialized in CFDate.c. In an effor to remedy Linux run loop support, __CFReadTSR (see ForFoundation.h) is now implemented on the POSIX RT clock_* APIs with a monotonic clock source which work at nanosecond granularity (though not necessarily nanosecond resolution).
Using the alernative "default" processor timebase solution had problems 
in that A) it may not work correctly in a multi-processor case and B) 
there is no standard API on Linux to get the timebase scalar and no 
arbitrary way to get it in a fashion consistent across Linux 
architectures (e.g. /proc/cpuinfo is different on i386 vs. powerpc). See 
CFDate.c for companion initialization code.