Ensure that the example works when the package is not installed or is not installed in a default location.
One of the use cases that the example runs through is validating the bundle for the package itself. However, since the package may or may not be installed at the time of the example execution and may or may not be installed in the default path, we install a temporary instance of the bundle using existing install rules, run the example, and then remove the temporary instance.