Added and leveraged 'kqueue_replace' for '__CFArmTimerInMode'.
'kqueue_replace' attempts to delete, if the event is present, the specified kqueue event and then replaces it through addition with an event of the same filter and
identifier using the specified flags, fflags, data, and udata.

The implementation and existence of 'kqueue_replace' is necessary to address the fact that, unlike the native in-kernel implementation of kqueues in BSD systems
such as FreeBSD or Darwin, libkqueue (at least as of 2.5.1) is unable to update the data portion of events such as EVFILT_TIMER. Consequently, the only way to set
the data expiry of such events is through the first-time add requiring this sort of delete-and-add replacement to work around this fact.

This is far less efficient than if we were able to call kqueue once and have it effect the input changelist as suggested by the man page and as implemented by BSD.